==> 00_Setup.Rmd <==
---
title: "00_Setup"
author: "Matt"
date: "10/23/2020"
output: html_document
---

Welcome to the setup script of for the Sechelt Cultural Plants project! Before proceeding, please ensure that you have the following programs installed on your machine:

1. R 4.0.0 or greater (32-bit required, 64-bit recommended, can have both installed at the same time. Available from [CRAN]("https://cran.r-project.org/"))
2. RStudio IDE for R (available from the [RStudio website]("https://rstudio.com/products/rstudio/download/#download"))

The intention of this script is to properly set up the computing environment and R packages required for carrying out this project. Checks and messages will appear at certain points if certain requirements are not met. Simply follow the instructions laid out to fix any issues you may be encountering.

```{r}

# Change default package installation option to prefer binary packages first
def_opt <- getOption("install.packages.compile.from.source")
options(install.packages.compile.from.source = "never")
cpus <- as.numeric(Sys.getenv("NUMBER_OF_PROCESSORS"))

# Some packages require devtools to be installed since they are not on CRAN and
# therefore need to be downloaded from github. Do that first
if(!"devtools" %in% installed.packages()[, "Package"])
  install.packages("devtools", Ncpus = cpus)

# First, check your system and your R version
if(.Platform$OS.type == "windows") {
  if(!"installr" %in% installed.packages()[, "Package"])
    install.packages("installr", Ncpus = cpus)
  if(as.numeric(version$major) < 4) {
    installr::install.R()
    stop(paste0("\r        
  ******************************************************************************
  R has now been upgraded to the latest version!
  With RStudio open, select \"Tools\" > \"Global Options\". In the \"General\" 
  tab, find the file path listing the current R version, and select the \"Change\"
  button to locate the installation path for the newly downloaded R version. 
  RStudio should populate a list of all currently installed R versions, but in
  case it does not your new R version is located here:\n
  ", normalizePath(tail(list.dirs(dirname(R.home()), recursive = FALSE), n = 1)), "
  
  You must now close and reopen RStudio in order for the changes to take effect.
  ******************************************************************************
  "))
  } else {
    message("R minimum requirements met")
  }
  
  # One of the packages that will be used below is raster, and it needs to be
  # the development version 3.3-14. The package requires compilation,
  # accomplished by rtools on Windows machines. Install rtools here if it
  # isn't already.
  if(!devtools::find_rtools()) {
    if(!"curl" %in% installed.packages()[, "Package"])
      install.packages("curl", Ncpus = cpus)
    if(.Platform$r_arch == "x64") {
      url <- "https://cran.r-project.org/bin/windows/Rtools/rtools40-x86_64.exe"
    } else {
      url <- "https://cran.r-project.org/bin/windows/Rtools/rtools40-i686.exe"
    }
    curl::curl_download(url, destfile = file.path(tempdir(), basename(url)))
    system(file.path(tempdir(), basename(url)))
    
    # Create environment variable
    writeLines('PATH="${RTOOLS40_HOME}\\usr\\bin;${PATH}"', con = "~/.Renviron")
    
    # Restart R
    stop("\r        
    ****************************************************************************
    You must now close and reopen RStudio in order for the changes to take effect
    ****************************************************************************
    "
    )
  } else {
    # Confirm that rtools is installed properly. If it didn't, follow the 
    # instructions on the rtools website: 
    # https://cran.r-project.org/bin/windows/Rtools/
    if(Sys.which("make") == "") {
      
      # Create environment variable
      writeLines('PATH="${RTOOLS40_HOME}\\usr\\bin;${PATH}"', con = "~/.Renviron")
      
      stop("rtools 4.0 is installed but not registered in your sytem PATH. An
      \rattempt has now been made to register rtools to the system PATH. Did
      \ryou close and reopen RStudio? If error persists after restarting, follow 
      \rthe instructions at https://cran.r-project.org/bin/windows/Rtools/")
    }
  }
} else {
  if(as.numeric(version$major) < 4) {
    stop("You must upgrade your R version to be at least version 4.0.0. Please
         \rvisit https://cran.r-project.org/ to download the latest version for
         \r your system")
  } else {
    message("R minimum requirements met")
  }
}

# Check that you are using RStudio
if(!"rstudioapi" %in% installed.packages()[, "Package"])
  install.packages("rstudioapi", Ncpus = cpus)
if(!rstudioapi::isAvailable()) {
  stop("RStudio is not running in this session. Have you installed it on your
       \rmachine? RStudio is available from https://rstudio.com/products/rstudio/download/#download")
} else {
  message("RStudio is running")
}

# One of the packages that will be used below is raster, and it needs to be the
# development version 3.3-14. The package requires compilation, accomplished by
# rtools on Windows machines. Install rtools here if it isn't already.

# List of packages used by all scripts
ls <- c("tidyverse", "raster", "sf", "terra", "tools", "bcdata", "bcmaps", "rgee", 
        "RStoolbox", "geojsonio", "googledrive", "stringi", "stars", "svSocket", 
        "RODBC", "apcluster", "ranger", "MLmetrics", "mlr3verse", "mlr3fselect", 
        "mlr3spatiotempcv", "corrplot", "GGally")

new_packages <- ls[!(ls %in% installed.packages()[, "Package"])]
if(length(new_packages)) 
  install.packages(new_packages, Ncpus = cpus)

# Standalone packages that require installation from github go here
remotes::install_github("mlr-org/mlr3extralearners")

# Required minimum versions for certain packages
if(compareVersion(packageDescription("raster")$Version, "3.3-14") < 0) 
  install.packages("rspatial/raster", Ncpus = cpus)
if(compareVersion(packageDescription("terra")$Version, "1.1-9") < 0) 
  remotes::install_github("rspatial/terra")
if(compareVersion(packageDescription("mlr3spatiotempcv")$Version, "0.0.0.9009") < 0) 
  install.packages("mlr-org/mlr3spatiotempcv", Ncpus = cpus)
if(compareVersion(packageDescription("mlr3")$Version, "0.11.0-9000") < 0)
  remotes::install_github("mlr-org/mlr3")

# Return default option
options(install.packages.compile.from.source = def_opt)

# Now that packages are installed, make sure a valid installation of SAGA GIS is
# downloaded to your PC. This is used in the first script to generate multiple
# terrain layers used in modelling later on
source(file.path("_functions", "get_saga.R"))
saga_cmd <- get_saga()
saga_ver <- system2(saga_cmd, "-v", stdout = TRUE)
saga_ver <- unlist(regmatches(saga_ver, regexec("Version:\\s*(.*?)\\s*\\(", saga_ver)))[2]

```
