==> 03_Model_Agreement.Rmd <==
---
title: "03_Model_Agreement"
author: "Matthew Coghill"
date: "7/14/2020"
output: html_document
---

The purpose of this file is to pull together all of the models run in the previous script and merge them together in order to find the areas where models are agreeing that a plant is present in a given raster cell. This is a fairly simple idea that was founded by Michael Ryan in our discussions. First, load the required packages.

```{r Load Packages, echo=TRUE, results='hide'}

invisible(suppressPackageStartupMessages(
  lapply(c("tidyverse", "terra"), library, character.only = TRUE)))

```

Next, load the directories that will be used for this script.

```{r Directory setup}

AOI <- "Sechelt"
AOI_dir <- file.path(".", paste0(AOI, "_AOI"))
map_res <- "TRIM"
res_dir <- ifelse(is.numeric(map_res), paste0(map_res, "m"), map_res)

model_dir <- file.path(AOI_dir, "2_map_predictions", res_dir)
covariate_dir <- file.path(AOI_dir, "1_map_inputs", "covariates", res_dir)

dsmart_dirs <- file.path(
  AOI_dir, "1_map_inputs", "dsmart_2019", res_dir, c("site_ser", "struct_stg"))

covariates <- rast(c(
  list.files(covariate_dir, pattern = ".tif$", full.names = TRUE),
  list.files(c(file.path(
    dsmart_dirs[endsWith(dsmart_dirs, "site_ser")], "simple"), 
    file.path(dsmart_dirs, "output", "probabilities")), 
    pattern = ".tif$", full.names = TRUE),
  list.files(dsmart_dirs, pattern = ".tif$", full.names = TRUE)))

```

Simple method proposed by Mike: Count the number of maps that predicted a plant at a given pixel. This requires a bit of adjustment as well: First, reclassify all map probability values below 0.5 to be 0, and then count the maps that predict anything above that value. Repeat for above and below 0.75.

```{r Pres Abs model agreement}

prob_in_dir <- file.path(model_dir, "pres_abs_predictions")
prob_out_dir <- file.path(
  AOI_dir, "3_model_agreement", res_dir, "pres_abs_agreement")
dir.create(prob_out_dir, showWarnings = FALSE, recursive = TRUE)
prob_cutoff <- c(0.5, 0.75) %>% magrittr::set_names(make.names(.))

out <- sapply(c("CORNCAN", "RUBUSPE"), function(i) {
  
  # Gather all of the model outputs
  model_stack <- rast(
    sapply(dir(prob_in_dir, pattern = i, full.names = TRUE), list.files, 
           pattern = "_TRUE.tif$", full.names = TRUE))
  
  mod_modified <- sapply(prob_cutoff, function(j) {
    # Classify cell values
    rcl <- classify(model_stack, matrix(
      c(0, j, 0, j, 1, 1), ncol = 3, byrow = TRUE))
    
    # Create summed raster
    agreement <- app(
      rcl, fun = "sum", na.rm = TRUE,
      filename = file.path(prob_out_dir, paste0(
        i, "_pres_abs_gt_", j * 100, "_model_agreement.tif")), 
      overwrite = TRUE, wopt = list(datatype = "INT2S"))
    
    agreement_freq <- as.data.frame(freq(agreement, bylayer = FALSE)) %>%
      magrittr::set_colnames(c("zone", "count"))
    
    # Zonal statistics
    z_stats <- c(sapply(c("mean", "min", "max", "sd"), function(k) {
      z <- zonal(covariates, agreement, fun = k, na.rm = TRUE)
      write.csv(z, file.path(out_dir, paste0(
        i, "_", k, "_pres_abs_gt_", j * 100, ".csv")), row.names = FALSE)
      return(z)}, simplify = FALSE, USE.NAMES = TRUE), 
      freq = list(agreement_freq))
    
    # Cleanup
    suppressWarnings(terra::tmpFiles(old = TRUE, remove = TRUE))
    
    return(z_stats)
  }, simplify = FALSE, USE.NAMES = TRUE)
}, simplify = FALSE, USE.NAMES = TRUE)

```

I thought of trying to do a similar technique with cover. Because of really small decimal places in raw cover values, I figured I should reclassify cover values less than 1% as a 0 value, and then everything else is a value of 1 in order to count where models agree that there is cover greater than 1%

```{r Cover model agreement}

# cov_in_dir <- file.path(model_dir, "cover_predictions")
# cov_out_dir <- file.path(AOI_dir, "3_model_agreement", res_dir, "cover_agreement")
# dir.create(cov_out_dir, showWarnings = FALSE, recursive = TRUE)
# cov_cutoff <- c(1, 5)
# 
# for(i in c("CORNCAN", "RUBUSPE")) {
#   model_stack <- rast(
#     sapply(dir(cov_in_dir, pattern = i, full.names = TRUE), list.files, 
#            pattern = "_cover.tif$", full.names = TRUE))
#   
#   for(j in cov_cutoff) {
#     rcl <- classify(model_stack, matrix(
#       c(0, j, 0, j, 100, 1), ncol = 3, byrow = TRUE))
#     
#     agreement <- app(
#       rcl, fun = "sum", na.rm = TRUE,
#       filename = file.path(cov_out_dir, paste0(
#         i, "_cover_gt_", j, "_model_agreement.tif")), 
#       overwrite = TRUE, wopt = list(datatype = "INT2S"))
#     
#     # Cleanup
#     suppressWarnings(terra::tmpFiles(old = TRUE, remove = TRUE))
#   }
# }

```

The outputs of the above calculations are "number of maps that agree that plant x has a probability of presence/has a cover percentage greater than y". These numbers of maps can be summed, and it's a weird metric but it is essentially the number of models that agree that a plant would be present and that the cover would be greater than 5%

```{r}

# full_out_dir <- file.path(AOI_dir, "3_model_agreement", res_dir, "full_agreement")
# dir.create(full_out_dir, showWarnings = FALSE, recursive = TRUE)
# 
# for(i in c("CORNCAN", "RUBUSPE")) {
#   for(j in cov_cutoff) {
#     for(k in prob_cutoff * 100) {
#       model_stack <- rast(c(
#         list.files(prob_out_dir, pattern = paste0(i, ".*.", k),
#                    full.names = TRUE), 
#         list.files(cov_out_dir, pattern = paste0(i, ".*.", j), 
#                    full.names = TRUE))) %>% 
#         magrittr::set_names(c("pres_abs", "cover"))
#       
#       # Only want to keep high cell values, indicating that many models agree
#       # at a given raster pixel
#       model_mask <- foreach(l = 1:nlyr(model_stack), .final = rast) %do% {
#         n_layers <- max(minmax(model_stack[[l]]))
#         classify(model_stack[[l]], matrix(
#           c(0, floor(n_layers / 2), 0, floor(n_layers / 2), Inf, 1), 
#           ncol = 3, byrow = TRUE))
#       } %>% 
#         sum() %>% 
#         classify(matrix(c(0, 2, 0, 2, Inf, 1), ncol = 3, byrow = TRUE))
#       
#       # Add both pres_abs and cover rasters, mask out areas defined from the 
#       # masking function above, and finally only keep cell values that are
#       # minimum of the mean of the number of agreeing models...
#       model_sum <- sum(model_stack, na.rm = TRUE) %>% 
#         mask(model_mask, maskvalue = 0, updatevalue = 0) %>% 
#         classify(
#           matrix(
#             c(0, floor(mean(minmax(model_stack)["rmax", ])), 0), 
#             ncol = 3, byrow = TRUE),
#           filename = file.path(
#             full_out_dir, paste0(i, "_cover_gt_", j, "_pres_abs_gt_",
#                               k, "_agreement.tif")), 
#           overwrite = TRUE, wopt = list(datatype = "INT2S"))
#     }
#   }
#   
#   # Produce a map of likely locations to find a given plant based on all other
#   # maps produced
#   out_maps <- rast(list.files(full_out_dir,
#     pattern = paste0(i, "_cover_.*..tif$"), full.names = TRUE))
#   likely_locations <- classify(out_maps, matrix(
#     c(0, 1, 0, 1, Inf, 1), ncol = 3, byrow = TRUE)) %>% 
#     sum() %>% 
#     classify(matrix(c(0, floor(nlyr(out_maps) / 2), 0), ncol = 3, byrow = TRUE), 
#              filename = file.path(full_out_dir, paste0(i, "_likely_locations.tif")), 
#              overwrite = TRUE, wopt = list(datatype = "INT2S"))
#   
#   # Cleanup
#   suppressWarnings(terra::tmpFiles(old = TRUE, remove = TRUE))
# }
# 
# # Attempt to find areas that all plants overlap based on the previous maps.
# likely_files <- list.files(
#   full_out_dir, pattern = "_likely_locations.tif", full.names = TRUE)
# efficiency <- rast(likely_files) %>% 
#   classify(matrix(c(0, 1, 0, 1, Inf, 1), ncol = 3, byrow = TRUE)) %>% 
#   sum() %>% 
#   classify(matrix(c(0, length(likely_files), 0, length(likely_files), Inf, 1),
#                   ncol = 3, byrow = TRUE), 
#            filename = file.path(full_out_dir, "efficient_locations.tif"), 
#            overwrite = TRUE, wopt = list(datatype = "INT2S"))
# 
# # Cleanup
# suppressWarnings(terra::tmpFiles(old = TRUE, remove = TRUE))

```
