==> 04a_PointData_Prep_Enhancement.Rmd <==
---
title: "04a_Model_Enhancement_DataPrep"
author: "Matt"
date: "8/25/2020"
output: html_document
---

In August 2020, another field visit to the Sechelt study area was organized and another week worth of field data was collected. New methods were used to collect this data and certain methods of how cover data was collected was even changed during the course of the field visit. This document is intended to walk through the steps of preparing the field data for model enhancement use. Essentially, data from the previous model runs will be combined with the data collected from this field visit and the models will be reran to see if the additional data helped with model accuracy. In this version, only presence/absence data will be used; however, cover data does exist, it is just not considered from here.

First, R packages need to be loaded. 

```{r Load Packages, echo=TRUE, results='hide'}

invisible(suppressPackageStartupMessages(
  lapply(c("tidyverse", "readxl", "sf", "terra"), 
         library, character.only = TRUE)))

```

Next, data from the field visits will be loaded in and processed. I'll first deal with site related data, and then move on to plant related data.

Since 2019, I've edited the original field data Excel spreadsheet. Instead of having separate sheets for each species of interest, they are now all combined on a single sheet. For each species, there is data for whether it was present or absent at a given site. Additionally, I've sifted through the data very carefully and where there were certain issues with the way it was laid out previously (e.g.: transition calls in the site series column, descriptions where it wasn't necessary for the structural stage and disturbance classes, etc.), I've edited them manually. This was done to maintain continuity between the newly collected field data and the data from 2019. In 2020, plant recordings only indicate that a plant was present, so the absence points for each site will need to be filled in here.

```{r Set directories}

AOI <- "Sechelt"
AOI_dir <- file.path(".", paste0(AOI, "_AOI"))
map_res <- c(4, 10, 25, "TRIM")
field_data_dir <- file.path(AOI_dir, "1_map_inputs", "field_data")

map_res_id <- c(
  paste0(suppressWarnings(map_res[!is.na(as.numeric(map_res))]), "m"),
  suppressWarnings(map_res[is.na(as.numeric(map_res))]))

```

The spreadsheets for site information for both years are loaded below. These will be used later on near the end.

```{r Field data 2019 2020 merge - Site}

data_2019 <- file.path(
  field_data_dir, "raw", 
  "shishalhBECandCulturalPlants_FieldData_2019_EDITS.xlsx")
data_2020 <- file.path(
  field_data_dir, "raw", "Sechelt_FieldSampling_Aug2020.xlsx")
sheets <- excel_sheets(data_2019)

bgc_2019 <- read_excel(data_2019, sheet = "sechelt_site") %>% 
  dplyr::rename(subzone = `BGC Subzone / Variant`, point_id = `Point ID`) %>% 
  stats::setNames(make.names(names(.))) %>% 
  dplyr::mutate(MapUnit = ifelse(is.na(Forested.Site.Series), Nonforested.Class,
                                 paste0(subzone, ".", Forested.Site.Series))) 
bgc_2020 <- read_excel(data_2020, sheet = "sechelt_site") %>% 
  dplyr::rename(subzone = `BGC Subzone / Variant`, point_id = `Point ID`) %>% 
  stats::setNames(make.names(names(.))) %>% 
  dplyr::mutate(MapUnit = ifelse(is.na(Forested.Site.Series), Nonforested.Class,
                                 paste0(subzone, ".", Forested.Site.Series))) 

```

This chunk will combine the two field data sets into a single plant data dataframe. For the 2019 data, fewer focal species were recorded after the first week of data collection, so the Excel sheet reflects the points where data was "*not recorded" for a given species at a site. For those records that indicate not being recorded, those records are removed from the final pres/abs dataset.

```{r Field data 2019 2020 merge - Plants}

# Plant species names in 2019 were concatenated, add space at the 4th character
# to be consistent with the 2020 naming
plants_2019 <- read_excel(data_2019, sheet = "sechelt_plants") %>% 
  merge(bgc_2019[, c("point_id", "MapUnit", "Lat", "Long")], by.x = "Point ID", by.y = "point_id") %>% 
  dplyr::mutate(Species = ifelse(nchar(Species) > 2, 
                                 gsub('(.{4})', '\\1 ', Species), Species)) %>% 
  dplyr::rename(point_id = `Point ID`)
plant_list_2019 <- unique(plants_2019$Species)

# In 2020, we only marked where plants were present at a site, not where they
# were absent. Need to expand the 2020 plant list to include absence points at
# each site. This differs from 2019 data collection where this was already
# filled in. Extra points that were opportunistically collected are appended
# since a whole site description wasn't filled out and they are simply just
# plant observations.
plants_2020 <- read_excel(data_2020, sheet = "sechelt_plants") %>% 
  merge(bgc_2020[, c("point_id", "MapUnit")], by.x = "Point ID", by.y = "point_id") %>% 
  dplyr::rename(point_id = `Point ID`)
plant_list_2020 <- unique(plants_2020$Species)

plants_2020_fix <- bind_rows(lapply(unique(bgc_2020$point_id), function(i) {
  site <- plants_2020 %>% 
    dplyr::filter(point_id == i) %>% 
    as.data.frame() %>% 
    dplyr::mutate(across(c(Lat, Long, A1, A2, A3, B1, B2, C, D), as.numeric))
  if(nrow(site) == 0) {
    site <- site %>% 
      add_row(point_id = i,
                    Lat = bgc_2020[bgc_2020$point_id == i, ]$Lat,
                    Long = bgc_2020[bgc_2020$point_id == i, ]$Long)
  }
  sp_add <- plant_list_2020[!plant_list_2020 %in% site$Species]
  
  out <- if(length(sp_add) > 0) {
    rbind(site, data.frame(
      point_id = i, 
      Lat = site$Lat[1], 
      Long = site$Long[1], 
      Species = sp_add, 
      A1 = NA_real_, A2 = NA_real_, A3 = NA_real_, 
      B1 = NA_real_, B2 = NA_real_, C = NA_real_, D = NA_real_,
      Flowers = NA_character_, Fruit = NA_character_, Comments = NA_character_) %>% 
        merge(dplyr::select(bgc_2020, point_id, MapUnit)) %>% 
        stats::setNames(gsub("\\.", " ", names(.)))) %>% 
      drop_na(Species)
  } else site
})) %>% rbind(dplyr::filter(plants_2020, !point_id %in% bgc_2020$point_id))

# Only consider points that had greater than 0.1% cover in a plot
plants_merge <- rbind(plants_2019, plants_2020_fix) %>% 
  dplyr::mutate(
    Cover = rowSums(.[, c("A1", "A2", "A3", "B1", "B2", "C", "D")], na.rm = TRUE),
    Pres = Cover > 0) %>% 
  stats::setNames(make.names(names(.))) %>% 
  dplyr::filter(!Comments %in% "*not recorded") 

# Use a list of dataframes for each plant species
plant_data <- lapply(unique(plants_merge$Species), function(x) {
  dplyr::filter(plants_merge, Species == x) %>% 
    dplyr::select(-Species)
}) %>% stats::setNames(unique(plants_merge$Species)) %>% 
  lapply(function(x) {
    st_as_sf(x, coords = c("Long", "Lat"), crs = 4326) %>% 
      st_transform(3005) %>% 
      dplyr::rename(geom = geometry)})

# Interesting tables of pres/abs frequencies
plant_table <- lapply(plant_data, function(x) data.frame(table(x$Pres)))

# If there is a skewed presence/absence ratio, remove the entire collection
# of data from the list. Arbitrary value of 5% given here. May want to remove this...
plant_data <- plant_data[sapply(plant_table, function(x) 
  if(min(x$Freq) / sum(x$Freq) >= 0.05) TRUE else FALSE)]

```

The next chunk will load in processed points from the 02a script, i.e.: unattributed points from TEM data collection as well as 2019 point data. The 2020 pres/abs points are added to the dataset.

```{r old point data}

old_pts <- sapply(map_res_id, function(z) {
  old_pto <- file.path(field_data_dir, "processed_2019", z)
  old_points <- lapply(c("RUBUSPE", "CORNCAN", "VERAVIR", "LEDUGRO"), function(x) {
    st_read(file.path(old_pto, "pres_abs_no_atts.gpkg"), layer = x, quiet = TRUE)
  }) %>% setNames(c("Rubu spe", "Corn can", "Vera vir", "Ledu gro"))
  
  # Need to filter away the 2019 processed points and keep just the points from 
  # TEM processing. 2019 points start with either M or Op, and end with _2019, so
  # use that pattern to filter out the appropriate points. Then, add in the 2019
  # and 2020 data
  
  points_comb <- c(sapply(
    names(old_points)[names(old_points) %in% names(plant_data)], function(x) {
      rbind(
        old_points[[x]] %>% 
          dplyr::filter(!grepl("^M|Op", point_id, ), !grepl("_2019$", point_id)), 
        plant_data[[x]])}, simplify = FALSE, USE.NAMES = TRUE), 
    plant_data[!names(plant_data) %in% names(old_points)], 
    old_points[!names(old_points) %in% names(plant_data)])
  return(points_comb)
}, simplify = FALSE, USE.NAMES = TRUE)

```

Now that the points are brought together from all data sources, the extraction from the rasters on them can once again be performed.

```{r Extract raster data}

pt_extract <- sapply(map_res_id, function(z) {
  covariate_dir <- file.path(AOI_dir, "1_map_inputs", "covariates", z)
  pto <- file.path(field_data_dir, "processed_2020", z)
  dir.create(pto, recursive = TRUE, showWarnings = FALSE)
  points_comb <- old_pts[[z]]
  
  invisible(sapply(names(points_comb), function(x) {
    points_comb[[x]] %>% 
      dplyr::select(point_id, MapUnit, Pres) %>% 
      cbind(st_coordinates(.)) %>% 
      st_write(file.path(pto, "pres_abs_bec_merge.gpkg"), layer = x, 
               delete_layer = TRUE, quiet = TRUE)}))
  
  # Create the raster stack (terra package) which will be used for data extraction
  covariates <- terra::rast(list.files(covariate_dir, full.names = TRUE, pattern = ".tif$"))
  
  pres_abs_2020 <- sapply(names(points_comb), function(x) {
    
    # Create unattributed points and write those
    st_write(points_comb[[x]], file.path(pto, "pres_abs_no_atts.gpkg"), layer = x, 
             delete_layer = TRUE, quiet = TRUE)
    
    # Attribute the points and write those separately
    out_att <- cbind(
      points_comb[[x]], 
      terra::extract(covariates, st_coordinates(points_comb[[x]]))) %>% 
      Filter(function(y) !all(is.na(y)), .) %>% 
      dplyr::select(Pres, names(covariates)) %>% 
      drop_na()
    st_write(out_att, file.path(pto, "pres_abs.gpkg"), layer = x, 
             delete_layer = TRUE, quiet = TRUE)
    return(out_att)
  }, simplify = FALSE, USE.NAMES = TRUE)
  return(pres_abs_2020)
}, simplify = FALSE, USE.NAMES = TRUE)

```

Lastly, we need to add the new site series and structural stage points from 2020 to the respective datasets. At the end, these points get saved to the processed_2020 folder.

```{r site series and structural stage points}

pt_extract <- sapply(map_res_id, function(z) {
  old_pto <- file.path(field_data_dir, "processed_2019", z)
  pto <- file.path(field_data_dir, "processed_2020", z)
  shapes_dir <- file.path(AOI_dir, "0_raw_inputs", "base_layers", z)
  
  # Load previous years site series data
  site_ser_old <- st_read(file.path(old_pto, "site_ser_pts.gpkg"), quiet = TRUE)
  
  aoi_sf <- st_read(file.path(shapes_dir, "aoi.gpkg"), quiet = TRUE) %>% 
    st_geometry() %>% st_make_valid()
  
  # Combine all site series data together
  site_ser_pts <- dplyr::select(
    bgc_2020, point_id, subzone, Forested.Site.Series,
    Nonforested.Class, Long, Lat) %>% 
    dplyr::mutate(across(c(Forested.Site.Series, Nonforested.Class), toupper)) %>% 
    dplyr::mutate(
      MapUnit = ifelse(
        is.na(Forested.Site.Series), Nonforested.Class, 
        paste0(subzone, ".", Forested.Site.Series))) %>% 
    st_as_sf(coords = c("Long", "Lat"), crs = 4326) %>% 
    st_transform(crs = st_crs(site_ser_old)) %>% 
    dplyr::rename(geom = geometry) %>% 
    dplyr::select(point_id, MapUnit) %>% 
    rbind(site_ser_old) %>% 
    st_set_agr("constant") %>% 
    st_intersection(aoi_sf)
  
  # Load previous years structural stage data
  struct_stg_old <- st_read(file.path(old_pto, "struct_stg_pts.gpkg"), 
                            quiet = TRUE) %>% 
    dplyr::select(StructuralStage)
  
  # Combine all structural stage data together
  struct_stg_pts <- dplyr::select(bgc_2020, Structural.Stage, Long, Lat) %>% 
    drop_na(Structural.Stage) %>% 
    st_as_sf(coords = c("Long", "Lat"), crs = 4326) %>% 
    st_transform(crs = st_crs(struct_stg_old)) %>% 
    dplyr::mutate(Structural.Stage = make.names(Structural.Stage)) %>% 
    dplyr::rename(StructuralStage = Structural.Stage, geom = geometry) %>% 
    rbind(struct_stg_old) %>% 
    st_set_agr("constant") %>% 
    st_intersection(aoi_sf)
  
  # Write outputs
  invisible(sapply(c("site_ser_pts", "struct_stg_pts"), function(x) st_write(
    eval(parse(text = x)), file.path(pto, paste0(x, ".gpkg")), 
    delete_layer = TRUE, quiet = TRUE)))
  return(list(site_series = site_ser_pts, struct_stg = struct_stg_pts))
}, simplify = FALSE, USE.NAMES = TRUE)

```
