==> 08b_Best_Covariates.Rmd <==
---
title: "08b_Best_Covariates"
author: "Matthew Coghill"
date: "2/8/2021"
output: html_document
---

The intent of this document is to use the best covariates from the previous modelling processes (the simple model runs) and create a model using those covariates. Much of this code will be copy/pasted from the simple model analysis file.

```{r}

invisible(suppressPackageStartupMessages(
  lapply(c("tidyverse", "terra", "sf", "mlr3verse", "corrplot", "ranger"), 
         library, character.only = TRUE, quietly = TRUE)))

# Load custom scripts
source("./_functions/model_gen_mlr3_binary_simple.R")

```

Next, define the locations and options for where files are located and how to continue processing.

```{r Define directories}

AOI <- "Sechelt"
AOI_dir <- file.path(".", paste0(AOI, "_AOI"))
map_res <- 4
res_dir <- ifelse(is.numeric(map_res), paste0(map_res, "m"), map_res)

input_dir <- file.path(
  "./Sechelt_AOI/1_map_inputs/field_data/processed_2020", res_dir)
covariate_dir <- file.path(AOI_dir, "1_map_inputs", "covariates", res_dir)
shapes_dir <- file.path(AOI_dir, "0_raw_inputs", "base_layers", res_dir)
dsmart_site_ser_dir <- file.path(AOI_dir, "1_map_inputs", "dsmart_2020", 
                                 res_dir, "site_ser")
dsmart_struct_stg_dir <- file.path(AOI_dir, "1_map_inputs", "dsmart_2020", 
                                   res_dir, "struct_stg")
output_dir <- file.path(AOI_dir, "8b_map_predictions_combined", res_dir)
dir.create(file.path(output_dir, "pres_abs_predictions"), showWarnings = FALSE)

```

For each plant species, go through their folders and find the variables that were most important

```{r}

simple_dir <- file.path(AOI_dir, "8_map_predictions_simple")
processed_res <- basename(list.dirs(simple_dir, recursive = FALSE))
simple_dir <- file.path(simple_dir, processed_res, "pres_abs_predictions")
sp_dirs <- list.dirs(simple_dir, recursive = FALSE)
best_vars <- bind_rows(lapply(sp_dirs, function(x) {
  best_models <- read.csv(file.path(x, paste0(basename(x), "_results.csv"))) %>% 
    dplyr::filter(!grepl("simple", cov_group))
  
  res <- basename(dirname(dirname(x)))
  
  # Get directory of best models
  bind_rows(lapply(best_models$cov_group, function(y) {
    z <- dplyr::filter(best_models, cov_group == y) %>% dplyr::pull(task_id)
    best_mod_dir <- file.path(x, y, z)
    read.csv(file.path(best_mod_dir, "variable_importance.csv")) %>% 
      dplyr::mutate(group = y)
  })) %>% dplyr::mutate(species = basename(x), res = res)
}))

best_vars <- list(`Corn can` = dplyr::filter(best_vars, species == "Corn can"),
                  `Rubu spe` = dplyr::filter(best_vars, species == "Rubu spe"))

var_counts <- lapply(best_vars, function(x) {
  # x <- best_vars[[1]]
  cnts <- x %>% 
    group_by(layer, group) %>% 
    summarise(n_models = n(), imp = sum(gini), .groups = "drop") %>% 
    dplyr::mutate(avg_imp = imp / n_models) %>% 
    group_by(group) %>% 
    dplyr::filter(n_models > 2, avg_imp >= 1) %>% 
    dplyr::mutate(n_group = n()) %>% 
    dplyr::arrange(desc(avg_imp))
  
  cnts2 <- dplyr::filter(cnts, n_group > 3, n_group < 10) %>% 
    dplyr::filter(avg_imp >= quantile(avg_imp, prob = 0.75))
  
  cnts3 <- dplyr::filter(cnts, n_group >= 10) %>% 
    dplyr::filter(avg_imp >= quantile(avg_imp, prob = 0.8))
  
  dplyr::filter(cnts, n_group <= 3) %>% 
    bind_rows(cnts2, cnts3)
})

vars_include <- lapply(var_counts, dplyr::pull, layer)
vars_include_chm <- if(map_res != "TRIM") {
  lapply(vars_include, c, "chm") %>% setNames(paste0(names(vars_include), "_w_chm"))
} else NULL
vars_include <- c(vars_include, vars_include_chm)
vars <- unique(unlist(vars_include))

```

Load the covariate layers and point data

```{r}

sat_layers <- grep("sentinel2_", vars, value = TRUE)
clim_layers <- grep("normal_", vars, value = TRUE)
dsmart_probs <- grep("dsmart_site_ser_probability", vars, value = TRUE)
dsmart_class <- grep("dsmart_site_ser_class", vars, value = TRUE)
terrain_layers <- vars[!vars %in% c(dsmart_probs, dsmart_class, sat_layers, clim_layers)]

# Special handling of dsmart classification because it is a factor
site_ser_dsmart_class <- rast(list.files(dsmart_site_ser_dir, pattern = "dsmart.tif$", full.names = TRUE)) %>% 
  setNames(dsmart_class) %>% 
  as.factor()

dsmart_class_labels <- data.frame(levels = levels(site_ser_dsmart_class)[[1]]$levels,
                                  code = levels(site_ser_dsmart_class)[[1]]$labels)

dsmart_class_lut <- read.table(file.path(dsmart_site_ser_dir, "output", "lookup.txt"), 
                               sep = ",", header = TRUE) %>% 
  merge(dsmart_class_labels) %>% 
  dplyr::select(levels, name)

levels(site_ser_dsmart_class) <- dsmart_class_lut

# Get CHM if it's there
chm <- if(map_res != "TRIM") {
  rast(file.path(covariate_dir, "chm.tif"))
} else NULL

# Need string to find files for DSMART probs
dsmart_find_string <- paste0(sub("dsmart_site_ser_probability", "", dsmart_probs), collapse = "|")

covariates <- c(
  rast(list.files(covariate_dir, full.names = TRUE, 
                  pattern = paste0(sat_layers, collapse = "|"))) %>% 
    setNames(sat_layers),
  rast(list.files(covariate_dir, full.names = TRUE, 
                  pattern = paste0(clim_layers, collapse = "|"))) %>% 
    setNames(clim_layers),
  rast(list.files(covariate_dir, full.names = TRUE, 
                  pattern = paste0(terrain_layers, collapse = "|"))) %>% 
    setNames(terrain_layers),
  rast(list.files(file.path(dsmart_site_ser_dir, "output", "probabilities"), 
             pattern = dsmart_find_string, full.names = TRUE)) %>% 
    setNames(dsmart_probs),
  site_ser_dsmart_class, 
  chm
)

# Read in attributed point data
pres_abs <- sapply(c("Corn can", "Rubu spe"), function(x) {
  st_read(file.path(input_dir, "pres_abs.gpkg"), layer = x, quiet = TRUE) %>% 
    mutate(across("Pres", as.factor)) %>%
    dplyr::select(Pres, names(covariates)[names(covariates) %in% names(.)])
  }, simplify = FALSE, USE.NAMES = TRUE)

# Determine missing covariates and create a SpatRaster of those
missing_covs <- subset(covariates, names(covariates)[!names(covariates) %in% names(pres_abs[[1]])])

# Extract missing data to pres_abs
pres_abs <- lapply(pres_abs, function(x) {
  x %>% cbind(terra::extract(missing_covs, st_coordinates(.)))})

```

Run the modelling algorithm

```{r}

models <- sapply(names(vars_include), function(j) {
  
  dir_out <- file.path(output_dir, "pres_abs_predictions", j)
  dir.create(dir_out, showWarnings = FALSE, recursive = TRUE)
  
  # Select the covariate group to use
  covariates_flt <- subset(covariates, vars_include[[j]])
  
  # Only select the point columns containing the covariate names, and omit rows
  # containing NA values
  sp <- sub("_w_chm", "", j)
  pres_abs_flt <- pres_abs[sp][[1]] %>% 
    dplyr::select(any_of(c("Pres", names(covariates_flt)))) %>% 
    drop_na()
  
  # traindat = pres_abs_flt
  # target = "Pres"
  # filter_correlation = TRUE
  # corr_cutoff = 0.95
  # positive = "TRUE"
  # id = j
  # covariates = covariates_flt
  # outDir = dir_out
  # folds = 10
  # repeats = 5
  # rfe = FALSE
  
  pres_model <- model_gen_mlr3_binary_simple(
    traindat = pres_abs_flt, target = "Pres",
    filter_correlation = TRUE, corr_cutoff = 0.95,
    positive = "TRUE", rfe = FALSE,
    id = j, covariates = covariates_flt,
    outDir = dir_out)
  
}, simplify = FALSE, USE.NAMES = TRUE)

```


```{r Results and map prediction}

results <- bind_rows(lapply(names(vars_include), function(x) {
  read.csv(file.path(output_dir, "pres_abs_predictions", x, "run_compare.csv")) %>% 
    dplyr::slice_max(mean_fbeta, n = max(2, nrow(.) * (1/3))) %>% 
    dplyr::slice_min(mean_bbrier, with_ties = FALSE) %>% 
    dplyr::mutate(species = x, .before = 1)
}))

sp <- names(pres_abs)

# Write results
for(i in sp) {
  invisible(dplyr::filter(results, grepl(i, species)) %>%
    write.csv(file.path(output_dir, "pres_abs_predictions",
                         paste0(i, "_results.csv")), row.names = FALSE))
}

# This can help filter the results if your R session crashed
# results <- results[c(7:nrow(results)), ]

# Run map predictions for best results
terraOptions(memfrac = 0.1)
best_preds <- rast(lapply(1:nrow(results), function(x) {
  sp <- results[x, ]$species
  run <- results[x, ]$task_id
  out <- file.path(output_dir, "pres_abs_predictions", sp)
  model <- readRDS(file.path(out, run, "best_model.rds"))
  covs <- subset(covariates, names(model$variable.importance))
  wopt <- list(names = "TRUE")
  pred <- terra::predict(
    covs, model, na.rm = TRUE,
    fun = function(model, ...) predict(model, verbose = FALSE, ...)$predictions,
    index = 1L,
    filename = file.path(out, paste0(wopt$names, ".tif")), overwrite = TRUE,
    wopt = wopt)
  # covs <- stack(subset(covariates[[cov_grp]], names(model$variable.importance)))
  # crs(covs) <- crs(covariates[[cov_grp]])
  # names(covs) <- names(subset(covariates[[cov_grp]], names(model$variable.importance)))
  # pred <- rast(raster::predict(
  #   covs, model, na.rm = TRUE,
  #   fun = function(model, ...) predict(model, verbose = FALSE, ...)$predictions,
  #   index = 1L), crs = crs(covariates[[cov_grp]])) %>%
  #   writeRaster(file.path(out, "TRUE.tif"), overwrite = TRUE, wopt = list(names = "TRUE"))
}))

```
